Move SourceFilesCollector from Finder to Filesystem
